Refactored HDHomeRun2.0 and DVBCaptureDevice's c code variable declaration for 32-bit 6-4bit compatible

+++ native/so/DVBCapture2.0/sage_DVBCaptureDevice.c
@@ -3030,7 +3030,7 @@ int  SageGetSatelliteTble( void* Capture, SAT_NAME *sat_name, int max_sat )
 	return 0;
 }
 
-void SageDelay( void* Capture, unsigned long ms )
+void SageDelay( void* Capture, uint32_t ms )
 {
  	usleep( ms );
 }
